Личные проекты/secaudit-platform/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "secaudit"
version = "0.1.0"
description = "Security Audit Platform (MVP)"
authors = [{name="You", email="you@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "PyYAML>=6.0.1",
    "jinja2>=3.1.4",
    "reportlab>=4.2.0",
]

[project.scripts]
secaudit = "secaudit.cli:main"
